requirements.txt
jira>=3.8.0
requests
python-dotenv
pandas